Improving documentation for generating the different datasets. Also replace bit.ly links with aka.ms ones
