+++ README.md
@@ -23,7 +23,7 @@ wget "https://bit.ly/2U3Mde2"
 ### Pre-training Graph Updater by Observation Generation
 ```
 # Download data for observation generation / contrastive observation classification
-cd obs_gen.0.1 ; wget https://bit.ly/3ep1yhI ; unzip obs_gen.0.1.zip ; cd ..
+cd obs_gen.0.1 ; wget https://aka.ms/twkg/obs_gen.0.1.zip ; unzip obs_gen.0.1.zip ; cd ..
 # Train
 python train_obs_generation.py configs/pretrain_observation_generation.yaml
 
@@ -32,7 +32,7 @@ python train_obs_generation.py configs/pretrain_observation_generation.yaml
 ### Pre-training Graph Updater by Contrastive Observation Classification
 ```
 # Download data for observation generation / contrastive observation classification
-cd obs_gen.0.1 ; wget https://bit.ly/3ep1yhI ; unzip obs_gen.0.1.zip ; cd ..
+cd obs_gen.0.1 ; wget https://aka.ms/twkg/obs_gen.0.1.zip ; unzip obs_gen.0.1.zip ; cd ..
 # Train
 python train_obs_infomax.py configs/pretrain_contrastive_observation_classification.yaml
 
@@ -41,7 +41,7 @@ python train_obs_infomax.py configs/pretrain_contrastive_observation_classificat
 ### Train Action Scorer with RL
 ```
 # Download games
-cd rl.0.2 ; wget https://bit.ly/2Mb4CBR ; unzip rl.0.2.zip ; cd ..
+cd rl.0.2 ; wget https://aka.ms/twkg/rl.0.2.zip ; unzip rl.0.2.zip ; cd ..
 # Modify configs/train_gata_rl.yaml
 #   L30: True to load pre-trained graph encoder, False to randomly initialize.
 #     L31:  'gata_pretrain_obs_gen_model', 'gata_pretrain_obs_infomax_model'. When L30 is True.
@@ -58,7 +58,7 @@ python train_rl_with_continuous_belief.py configs/train_gata_rl.yaml
 ### Pre-training Graph Encoder by Action Prediction
 ```
 # Download data
-cd ap.0.2 ; wget https://bit.ly/2v6nbC1 ; unzip ap.0.2.zip ; cd ..
+cd ap.0.2 ; wget https://aka.ms/twkg/ap.0.2.zip ; unzip ap.0.2.zip ; cd ..
 # Train
 python train_action_prediction.py configs/pretrain_action_prediction_full.yaml
 
@@ -67,7 +67,7 @@ python train_action_prediction.py configs/pretrain_action_prediction_full.yaml
 ### Pre-training Graph Encoder by State Prediction
 ```
 # Download data
-cd sp.0.2 ; wget https://bit.ly/2Uyj9wS ; unzip sp.0.2.zip ; cd ..
+cd sp.0.2 ; wget https://aka.ms/twkg/sp.0.2.zip ; unzip sp.0.2.zip ; cd ..
 # Train
 python train_state_prediction.py configs/pretrain_state_prediction_full.yaml
 
@@ -76,7 +76,7 @@ python train_state_prediction.py configs/pretrain_state_prediction_full.yaml
 ### Pre-training Graph Encoder by Deep Graph Infomax
 ```
 # Download data
-cd dgi.0.2 ; wget https://bit.ly/383vAEQ ; unzip dgi.0.2.zip ; cd ..
+cd dgi.0.2 ; wget https://aka.ms/twkg/dgi.0.2.zip ; unzip dgi.0.2.zip ; cd ..
 # Train
 python train_deep_graph_infomax.py configs/pretrain_deep_graph_infomax_full.yaml
 
@@ -85,7 +85,7 @@ python train_deep_graph_infomax.py configs/pretrain_deep_graph_infomax_full.yaml
 ### Train Action Scorer with RL
 ```
 # Download games
-cd rl.0.2 ; wget https://bit.ly/2Mb4CBR ; unzip rl.0.2.zip ; cd ..
+cd rl.0.2 ; wget https://aka.ms/twkg/rl.0.2.zip ; unzip rl.0.2.zip ; cd ..
 # Modify configs/train_gata_gtf_rl.yaml
 #   L30: True to load pre-trained graph encoder, False to randomly initialize.
 #     L31:  'gata_gtf_pretrain_ap_full_model', 'gata_gtf_pretrain_sp_full_model', or 'gata_gtf_pretrain_dgi_full_model'. When L30 is True.
@@ -101,7 +101,7 @@ python train_rl_with_ground_truth_discrete_belief.py configs/train_gata_gtf_rl.y
 ### Pre-training Graph Encoder by Action Prediction
 ```
 # Download data
-cd ap.0.2 ; wget https://bit.ly/2v6nbC1 ; unzip ap.0.2.zip ; cd ..
+cd ap.0.2 ; wget https://aka.ms/twkg/ap.0.2.zip ; unzip ap.0.2.zip ; cd ..
 # Train
 python train_action_prediction.py configs/pretrain_action_prediction_seen.yaml
 
@@ -110,7 +110,7 @@ python train_action_prediction.py configs/pretrain_action_prediction_seen.yaml
 ### Pre-training Graph Encoder by State Prediction
 ```
 # Download data
-cd sp.0.2 ; wget https://bit.ly/2Uyj9wS ; unzip sp.0.2.zip ; cd ..
+cd sp.0.2 ; wget https://aka.ms/twkg/sp.0.2.zip ; unzip sp.0.2.zip ; cd ..
 # Train
 python train_state_prediction.py configs/pretrain_state_prediction_seen.yaml
 
@@ -119,7 +119,7 @@ python train_state_prediction.py configs/pretrain_state_prediction_seen.yaml
 ### Pre-training Graph Encoder by Deep Graph Infomax
 ```
 # Download data
-cd dgi.0.2 ; wget https://bit.ly/383vAEQ ; unzip dgi.0.2.zip ; cd ..
+cd dgi.0.2 ; wget https://aka.ms/twkg/dgi.0.2.zip ; unzip dgi.0.2.zip ; cd ..
 # Train
 python train_deep_graph_infomax.py configs/pretrain_deep_graph_infomax_seen.yaml
 
@@ -128,7 +128,7 @@ python train_deep_graph_infomax.py configs/pretrain_deep_graph_infomax_seen.yaml
 ### Pre-training Graph Updater by Command Generation
 ```
 # Download data for command generation
-cd cmd_gen.0.2 ; wget https://bit.ly/385UVOC ; unzip cmd_gen.0.2.zip ; cd ..
+cd cmd_gen.0.2 ; wget https://aka.ms/twkg/cmd_gen.0.2.zip ; unzip cmd_gen.0.2.zip ; cd ..
 # Train
 python train_command_generation.py configs/pretrain_command_generation.yaml
 
@@ -137,7 +137,7 @@ python train_command_generation.py configs/pretrain_command_generation.yaml
 ### Train Action Scorer with RL
 ```
 # Download games
-cd rl.0.2 ; wget https://bit.ly/2Mb4CBR ; unzip rl.0.2.zip ; cd ..
+cd rl.0.2 ; wget https://aka.ms/twkg/rl.0.2.zip ; unzip rl.0.2.zip ; cd ..
 # Modify configs/train_gata_gtp_rl.yaml
 #   L30: True to load pre-trained graph encoder, False to randomly initialize.
 #     L31:  'gata_gtp_pretrain_ap_seen_model', 'gata_gtp_pretrain_sp_seen_model', or 'gata_gtp_pretrain_dgi_seen_model'. When L30 is True.
