+++ dgi.0.2/README.md
@@ -1,12 +1,22 @@
-# Make sure you are in the root folder of the project.
-# Requirement: collect playthroughs first (see ./playthroughs/README.md).
-# wget https://bit.ly/383vAEQ
+# Building the datasets needed for pre-training the graph encoder by Deep Graph Infomax
 
-PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/train.json --out ./dgi.0.2/train.seen.json --graph-type seen
-PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/train.json --out ./dgi.0.2/train.full.json --graph-type full
+GATA's graph encoder can be pre-trained using Deep Graph Infomax. We make use of trajectories extracted from playing several text-based games.
 
-PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/valid.json --out ./dgi.0.2/valid.seen.json --graph-type seen
-PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/valid.json --out ./dgi.0.2/valid.full.json --graph-type full
+For convenience, the datasets were made available at [https://aka.ms/twkg/dgi.0.2.zip](https://aka.ms/twkg/dgi.0.2.zip). Once downloaded, extract its contents into `dgi.0.2/`.
 
-PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/test.json --out ./dgi.0.2/test.seen.json --graph-type seen
-PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/test.json --out ./dgi.0.2/test.full.json --graph-type full
+## Prerequisite
+
+Collect playthroughs (see [../playthroughs/README.md](../playthroughs/README.md)).
+
+## Instructions
+
+Make sure you are in the root folder of the project. Then run,
+
+    PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/train.json --out ./dgi.0.2/train.seen.json --graph-type seen
+    PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/train.json --out ./dgi.0.2/train.full.json --graph-type full
+
+    PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/valid.json --out ./dgi.0.2/valid.seen.json --graph-type seen
+    PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/valid.json --out ./dgi.0.2/valid.full.json --graph-type full
+
+    PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/test.json --out ./dgi.0.2/test.seen.json --graph-type seen
+    PYTHONPATH=. python ./dgi.0.2/make_dgi_dataset.py ./playthroughs/test.json --out ./dgi.0.2/test.full.json --graph-type full
