+++ playthroughs/collect_playthroughs.py
@@ -1,9 +1,6 @@
-import io
 import os
 import json
-import zipfile
 import argparse
-import urllib.request
 import multiprocessing
 from os.path import join as pjoin
 
@@ -17,56 +14,6 @@ from textworld.logic import State, Rule, Proposition, Variable
 from generic import process_facts, serialize_facts, gen_graph_commands
 from generic import process_local_obs_facts, process_fully_obs_facts
 from viz import show_kg
-
-
-ZIP_FILENAME = "TextWorld_CoG2019.zip"
-GAMES_URL = "https://aka.ms/ftwp/dataset.zip"
-
-
-def download(url, filename=None, force=False):
-    filename = filename or url.split('/')[-1]
-
-    if os.path.isfile(filename) and not force:
-        return filename
-
-    def _report_download_status(chunk_id, max_chunk_size, total_size):
-        size = chunk_id * max_chunk_size / 1024**2
-        size_total = total_size / 1024**2
-        unit = "Mb"
-        if size <= size_total:
-            print("{:.1f}{unit} / {:.1f}{unit}".format(size, size_total, unit=unit), end="\r")
-
-    filename, _ = urllib.request.urlretrieve(url, filename, _report_download_status)
-    return filename
-
-
-def extract_games(zip_filename, dst):
-    zipped_file = zipfile.ZipFile(zip_filename)
-    filenames_to_extract = [f for f in zipped_file.namelist() if f.endswith(".ulx") or f.endswith(".json")]
-
-    subdirs = {
-        "train": pjoin(dst, "train"),
-        "valid": pjoin(dst, "valid"),
-        "test": pjoin(dst, "test"),
-    }
-    for d in subdirs.values():
-        if not os.path.isdir(d):
-            os.makedirs(d)
-
-    print("Extracting...")
-    extracted_files = []
-    for filename in tqdm.tqdm(filenames_to_extract):
-        subdir = subdirs[os.path.basename(os.path.dirname(filename))]
-        out_file = pjoin(subdir, os.path.basename(filename))
-        extracted_files.append(out_file)
-        if os.path.isfile(out_file):
-            continue
-
-        data = zipped_file.read(filename)
-        with open(out_file, "wb") as f:
-            f.write(data)
-
-    return extracted_files
 
 
 def collect_data_from_game(gamefile):
@@ -154,8 +101,7 @@ def collect_data(gamefiles, args):
     if args.nb_processes > 1:
         pool = multiprocessing.Pool(args.nb_processes)
         results = []
-        for i, gamefile in enumerate(gamefiles):
-            seed = args.seed + i
+        for gamefile in gamefiles:
             result = pool.apply_async(collect_data_from_game, (gamefile,), callback=_assemble_results)
             results.append(result)
 
@@ -167,7 +113,6 @@ def collect_data(gamefiles, args):
 
     else:
         for i, gamefile in enumerate(gamefiles):
-            seed = args.seed + i
             data = collect_data_from_game(gamefile)
             _assemble_results(data)
 
@@ -186,13 +131,9 @@ def build_argparser():
     parser.add_argument("--nb-processes", type=int,
                         help="Number of CPUs to use. Default: all available.")
 
-    parser.add_argument("--seed", type=int, default=20190423,
-                        help="Seed for the random exploration.")
-    parser.add_argument("--branching-depth", type=int, default=0,
-                        help="Number of random commands for each transition in the walkthrough. Default: %(default)s.")
+    parser.add_argument("games", nargs="+",
+                        help="List of games (or folders containing games) to collect the playthroughs from.")
 
-    parser.add_argument("--games-dir", default="./games/",
-                        help="Folder where to extract the downloaded games.")
     parser.add_argument("-f", "--force", action="store_true",
                         help="Overwrite existing files.")
 
@@ -207,12 +148,14 @@ def main():
     if os.path.isfile(args.output) and not args.force:
         parser.error("{} already exists. Use -f to overwrite.".format(args.output))
 
-    if not os.path.exists(args.games_dir):
-        filename = download(GAMES_URL, filename=ZIP_FILENAME, force=args.force)
-        extracted_files = extract_games(filename, dst=args.games_dir)
-        gamefiles = [f for f in extracted_files if f.endswith(".json")]
-    else:
-        gamefiles = [args.games_dir + f for f in os.listdir(args.games_dir) if f.endswith(".json")]
+    gamefiles = []
+    for e in args.games:
+        if os.path.isdir(e):
+            gamefiles += [pjoin(e, f) for f in os.listdir(e) if f.endswith(".json")]
+        elif e.endswith(".json"):
+            gamefiles.append(e)
+        else:
+            print("Ignoring non-JSON files '{}'".format(e))
 
     collect_data(gamefiles, args)
